Update calculations and coins

+++ core/src/main/kotlin/me/kpavlov/elven/ai/Calculator.kt
@@ -2,6 +2,7 @@ package me.kpavlov.elven.ai
 
 import dev.langchain4j.agent.tool.Tool
 import ktx.log.logger
+import java.math.BigDecimal
 
 private val logger = logger<Calculator>()
 
@@ -15,24 +16,24 @@ internal object Calculator {
 
     @Tool("Calculates the sum of two numbers")
     fun add(
-        a: Double,
-        b: Double,
-    ): Double {
+        a: BigDecimal,
+        b: BigDecimal,
+    ): BigDecimal {
         logger.info { "🧮 Called add with a=$a, b=$b" }
         return a + b
     }
 
     @Tool("Calculates the square root of a number")
-    fun sqrt(x: Int): Double {
+    fun sqrt(x: BigDecimal): Double {
         logger.info { "🧮 Called sqrt with x=$x" }
         return kotlin.math.sqrt(x.toDouble())
     }
 
     @Tool("Returns the result of multiplication of two integers")
     fun multiply(
-        a: Double,
-        b: Double,
-    ): Double {
+        a: BigDecimal,
+        b: BigDecimal,
+    ): BigDecimal {
         logger.info { "🧮 Multiplying $a and $b" }
         return a * b
     }
